fixing rust warnings issues
