+++ contracts/fpt-staking-contract/tests/failure.rs
@@ -24,7 +24,9 @@ async fn fails_unstake_wrong_amount() {
     .await;
     let mock_token_asset_id = mock_token.contract_id().asset_id(&AssetId::zeroed().into());
 
-    fpt_staking_abi::stake(&contracts.fpt_staking, mock_token_asset_id, 1 * PRECISION).await;
+    fpt_staking_abi::stake(&contracts.fpt_staking, mock_token_asset_id, 1 * PRECISION)
+        .await
+        .unwrap();
 
     fpt_staking_abi::unstake(
         &contracts.fpt_staking,
