+++ deploy-scripts/Cargo.toml
@@ -15,7 +15,8 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 test-utils = { workspace = true }
 tokio = { workspace = true }
-
+csv = { workspace = true }
+tai64 = { workspace = true }
 
 [lib]
 doctest = false
