+++ test-utils/src/interfaces/vesting.rs
@@ -1,7 +1,6 @@
-use csv::ReaderBuilder;
 use fuels::prelude::{Address, Bech32Address};
 use fuels::types::errors::Error;
-use fuels::types::{AssetId, ContractId};
+use fuels::types::AssetId;
 use fuels::{
     prelude::{abigen, Account},
     programs::responses::CallResponse,
@@ -18,7 +17,7 @@ abigen!(Contract(
     abi = "contracts/vesting-contract/out/debug/vesting-contract-abi.json"
 ));
 
-const TOTAL_AMOUNT_VESTED: u64 = 100_000_000 * 68 / 100 * PRECISION;
+pub const TOTAL_AMOUNT_VESTED: u64 = 100_000_000 * 68 / 100 * PRECISION;
 
 pub async fn instantiate_vesting_contract<T: Account>(
     contract: &VestingContract<T>,
@@ -101,84 +100,4 @@ struct IntermediateVestingSchedule {
     claimed_amount: u64,
     total_amount: u64,
     recipient: String, // Assume the JSON contains a string for the recipient.
-}
-
-pub fn load_vesting_schedules_from_csv(
-    path: &str,
-    cliff_percentage: f64,
-    seconds_to_cliff: u64,
-    seconds_vesting_duration: u64,
-) -> Vec<VestingSchedule> {
-    let absolute_path = get_absolute_path_from_relative(path);
-    let file = File::open(&absolute_path).expect("Failed to open file");
-    let mut reader = ReaderBuilder::new()
-        .flexible(true)
-        .trim(csv::Trim::All)
-        .from_reader(file);
-
-    let now_unix = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .expect("Time went backwards")
-        .as_secs();
-
-    let cliff_timestamp = now_unix + seconds_to_cliff;
-    let end_timestamp = cliff_timestamp + seconds_vesting_duration;
-
-    let cliff_timestamp = tai64::Tai64::from_unix(cliff_timestamp.try_into().unwrap());
-    let end_timestamp = tai64::Tai64::from_unix(end_timestamp.try_into().unwrap());
-
-    let mut schedules = Vec::new();
-
-    for result in reader.records() {
-        let record = result.expect("Failed to read CSV record");
-        if record.len() < 5 || record[1].is_empty() {
-            continue; // Skip empty or invalid rows
-        }
-
-        println!("record: {:?}", record);
-
-        let total_amount =
-            (record[1].replace([',', '"'], "").parse::<f64>().unwrap() * PRECISION as f64) as u64;
-        let recipient = if !record[2].is_empty() {
-            Identity::Address(Address::from_str(&record[2]).unwrap())
-        } else if !record[3].is_empty() {
-            continue;
-            // ignore the recipient for now since ETH addresses are not supported yet
-            // Identity::Address(Address::from_str(&record[3]).unwrap())
-        } else {
-            continue; // Skip if both wallet addresses are empty
-        };
-
-        let schedule = VestingSchedule {
-            cliff_amount: (total_amount as f64 * cliff_percentage) as u64,
-            cliff_timestamp: cliff_timestamp.0,
-            end_timestamp: end_timestamp.0,
-            claimed_amount: 0,
-            total_amount,
-            recipient,
-        };
-
-        println!("schedule: {:?}", schedule);
-
-        schedules.push(schedule);
-    }
-    // take the sum of all total_amounts
-    let total_sum: u64 = schedules.iter().map(|s| s.total_amount).sum();
-    println!("Total sum of all vesting amounts: {}", total_sum);
-    // add one more schedule with the remaining amount
-    // TODO: find the total amount vested
-    schedules.push(VestingSchedule {
-        cliff_amount: total_sum,
-        cliff_timestamp: cliff_timestamp.0,
-        end_timestamp: end_timestamp.0,
-        claimed_amount: 0,
-        total_amount: TOTAL_AMOUNT_VESTED - total_sum,
-        recipient: Identity::ContractId(
-            ContractId::from_str(
-                "0x4761863a5b9a7ec3263964f694f453a5a67cf0d458ebc3e36eb618d43809c785",
-            )
-            .unwrap(),
-        ),
-    });
-    schedules
 }
